# [Merged by Bors] - fix(ring_theory/ideal/basic): ideal.module_pi speedup