Eric and Yael were both complaining that `ideal.module_pi` would occasionally cause random timeouts on unrelated PRs. This PR (a) makes the `smul` proof obligation much tidier (factoring out a sublemma) and (b) replaces the `all_goals` trick by 6 slightly more refined proofs (making the new proof longer, but quicker). On my machine the profiler stats are:

```
ORIG

parsing took 74.1ms
elaboration of module_pi took 3.83s
type checking of module_pi took 424ms
decl post-processing of module_pi took 402ms

NEW

parsing took 136ms
elaboration of module_pi took 1.19s
type checking of module_pi took 82.8ms
decl post-processing of module_pi took 82.5ms
```

---
<!-- The text above the `---` will become the commit message when your
PR is merged. Please leave a blank newline before the `---`, otherwise
GitHub will format the text above it as a title.

Any other comments you want to keep out of the PR commit should go
below the `---`, and placed outside this HTML comment, or else they
will be invisible to reviewers.

If this PR depends on other PRs, please list them below this comment,
using the following format:
- [ ] depends on: #abc [optional extra text]
- [ ] depends on: #xyz [optional extra text]
-->

Eric/Yael conversation is here: https://leanprover.zulipchat.com/#narrow/stream/113488-general/topic/distrib_mul_action_with_zero/near/252816196

I'm sorry the proof is now longer, but if this is one of these proofs which occasionally tips CI over the edge then it's probably best to have it much quicker and a bit longer. To orientate you around the diff: the smul proof is now a third the size and much less gnarly; it now finishes on line 644 not 647 even though we have some new lemma `mul_sub_mul_mem` before that, so this is a win. Github diff gets a bit confused about what happens next -- the main point is that I removed `all_goals {generic proof}` on line 648 and replaced it with six similar bespoke proofs for the remaining obligations. Just refactoring the smul proof and leaving the `all_goals` technique in there makes elaboration on my machine go back up to 1.6 seconds; this would be an alternative approach, and if we did this instead then the proof would be both shorter and quicker, but not as quick as this version.


[![Open in Gitpod](https://gitpod.io/button/open-in-gitpod.svg)](https://gitpod.io/from-referrer/)